Make the RockerCompile task incremental.
This commit makes rocker template generation incremental, meaning that
when a template file is changed we regenerate only that file (or files)
instead of running the Rocker compiler against all existing templates
again. Stale templates are also cleaned as expected, so that when a
template is removed its corresponding Java source file is deleted
without having to rerun the Rocker compiler.