Remove empty directories from the generated output directory
In case of incremental compilation, the removed templates are correctly  triggering a deletion of the corresponding compiled templates.
Yet, empty directories could remain in the generated output directory.
If this directory is used as an input to a subsequent Gradle task, the empty directories will cause a different cache key computation (compared to a case where the same rocker templates are declared, but without these empty directories in the generated directory), causing a cache miss for this subsequent task.